File upload
a developer deleted the directories due to an apparent security issue, though we have not seen any. so please aid caution

the developer has had access revoked. we also uploaded the wrong files in a previous commit, files from the developer's GH
apparently this will not work forever, so enjoy it while it lasts